User registration

+++ base/views.py
@@ -5,9 +5,13 @@ from .forms import RoomForm
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib. auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method == 'post':
         username = request.POST.get('username')
         password = request.POST.get('username')
@@ -51,7 +55,7 @@ def room(request,pk):
     context = {'room' : room}
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -64,10 +68,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('you are not allowed here')
+
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -78,8 +86,13 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+
+@login_required(login_url='login')
 def deleteroom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('you are not allowed here')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
